Add special support for Integer, Float and String
Creates a new data struct in the prelude library for Integer Float
and String types. These types require some special support from the
compiler at the moment because there isn't a way to represent their
implementation in language at the moment. This will hopefully change
in the future, but for now we perform a special check when searching
for implementations for implicit values and if it matches one of
these types we essentially just return true.