Remove stateful unique ids
Previously, many unique ids that were used for generated variable names
had some state that was retained over multiple compilations if the node
process did not exit (eg. in tests). This PR removes most of that state
so that ids start from 1 in each test. There is still one persistent
id usage in the constructors file which I left because I didn't think
that it would stick around for much longer.