Fix bug when typing applications as params
Previously, when there is an application in a function parameter
and the callee is a free variable, the type of the parameter would
also end up being a free variable which is wrong because we loose
all type information about the parameter. Instead, I added a
somewhat hacky check when typing applications that doesn't simplify
applications if the callee is a free variable and instead returns
the application un-reduced.